Whitespace handling and 2 decimal point format
Adding whitespace caused the code before to fire the else condition, prompting to the user that "*blank* is not a valid amount". Also, the float functions caused the total amount to sometimes display 13 decimal points to the right. Added an elif condition to skip the line when spaces are found. Then had the match case block multiply by 100 to get rid of the decimal point for flawless adding/subtracting. And then included a line to divide by 100 to bring back the decimal point for the total amount to be formatted by 2 decimal points.